ReportsDetail: Reduce number of functions
